refactor:  separated classes HttpSocket and HttpsSocket into header and implementation files
